[Fix] MapDataParser 관련 버그 고침

+++ src/main/java/org/newdawn/spaceinvaders/map_load/map_load_commands/InstantiateCommand.java
@@ -6,21 +6,21 @@ import org.newdawn.spaceinvaders.enums.GameObjectType;
 
 public class InstantiateCommand extends MapLoadCommand {
     private long _instantiateTime;
-    private int _instantiateX;
-    private int _instantiateY;
+    private long _instantiateX;
+    private long _instantiateY;
     private GameObjectType _gameObjectType;
     private String _gameObjectName;
     private ArrayList<String> _extra;
 
     public long getInstantiateTime() { return _instantiateTime; }
-    public int getInstantiateX() { return _instantiateX; }
-    public int getInstantiateY() { return _instantiateY; }
+    public long getInstantiateX() { return _instantiateX; }
+    public long getInstantiateY() { return _instantiateY; }
     public GameObjectType getGameObjectType() { return _gameObjectType; }
     public String getGameObjectName() { return _gameObjectName; }
     public ArrayList<String> getExtra() { return _extra; }
 
     public InstantiateCommand(
-        long instantiateTime, int instantiateX, int instantiateY, 
+        long instantiateTime, long instantiateX, long instantiateY, 
         GameObjectType gameObjectType, String gmaeObjectName,
         ArrayList<String> extra)
     {
